rename test files
+ move auth test to tests_auth.py
